core: First pass of platform event abstraction
This is the first in a series of changes that will ultimately provide
complete abstraction of event sources. The core library will no longer
require any knowledge of which platform it is being run on and which
resources (e.g. events, timers) are available.

This change defines the basic abstraction types and functions. Most
existing code that was written specifically for file descriptors has
been either renamed if leveragable or removed if not platform agnostic.
Note that the API has been completely preserved.

The actual abstraction implementation is not yet defined for any platform
with this change. All functions are merely stubs at this point.

Signed-off-by: Chris Dickens <christopher.a.dickens@gmail.com>